+++ client/src/models/sudoku_logic/square_constructor.js
